requirements.txt
discord-py-interactions
aiocache[redis]
python-dotenv
motor